requirements.txt
# LangChain packages - compatible versions
langchain==0.1.20
langchain-community==0.0.38
langchain-text-splitters==0.0.1
langchain-core==0.1.52
langchain-groq==0.1.3

# Vector Store & Embeddings
faiss-cpu==1.8.0
sentence-transformers==2.3.1

# Document Processing
pypdf==3.17.4
python-dotenv==1.0.0

# FastAPI (for web interface)
fastapi==0.109.0
uvicorn[standard]==0.27.0
websockets==12.0
python-multipart==0.0.6

# Utilities (optional - remove if causing issues)
tiktoken==0.5.2
transformers==4.36.2

# Note: torch removed - sentence-transformers will install compatible version